+++ protocols/kad/tests/client_mode.rs
@@ -59,11 +59,11 @@ async fn two_servers_add_each_other_to_routing_table() {
 
     match libp2p_swarm_test::drive(&mut server2, &mut server1).await {
         (
-            [Identify(_), Kad(UnroutablePeer { .. }), Identify(_), Kad(RoutingUpdated { peer: peer2, .. }), Identify(_)],
+            [Identify(_), Kad(RoutingUpdated { peer: peer2, .. }), Identify(_)],
             [Identify(_), Identify(_)],
         )
         | (
-            [Identify(_), Kad(UnroutablePeer { .. }), Identify(_), Identify(_), Kad(RoutingUpdated { peer: peer2, .. })],
+            [Identify(_), Identify(_), Kad(RoutingUpdated { peer: peer2, .. })],
             [Identify(_), Identify(_)],
         ) => {
             assert_eq!(peer2, server1_peer_id);