+++ protocols/kad/CHANGELOG.md
@@ -1,9 +1,12 @@
-## 0.44.4 - unreleased
+## 0.44.4
 
 - Implement common traits on `RoutingUpdate`.
   See [PR 4270].
+- Reduce noise of "remote supports our protocol" log.
+  See [PR 4278].
 
 [PR 4270]: https://github.com/libp2p/rust-libp2p/pull/4270
+[PR 4278]: https://github.com/libp2p/rust-libp2p/pull/4278
 
 ## 0.44.3
 
@@ -11,6 +14,11 @@
   See [PR 4224].
 
 [PR 4224]: https://github.com/libp2p/rust-libp2p/pull/4224
+
+- Rename missed `KademliaEvent::OutboundQueryCompleted` to `KademliaEvent::OutboundQueryProgressed` in documentation.
+  See [PR 4257].
+
+[PR 4257]: https://github.com/libp2p/rust-libp2p/pull/4257
 
 ## 0.44.2